Switched NotebookPicker to navbar
