+++ src/components/navbar.css
@@ -1,0 +1,12 @@
+.parent {
+  height: 75vh;
+  width: 75vw;
+  position: relative;
+}
+
+.spinner {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
